feat(authup,server-core)!: the config command lives with the CLI
`authup config schema` prints the whole `authup.yml` document, so the docs
workflow builds the CLI and writes `docs/src/public/schema/config.json`
before it builds the site, and triggers on the config sources as well as on
`docs/**`. The committed copy and server-core's `dist/config-schema.json`
are gone with the script that emitted them: both were a copy a registry
change could leave stale, and that script imported a module the config
rework had already deleted, so the server build was failing.

`config validate` moves out of server-core into `apps/authup` with it, since
both subcommands are about the whole document and the whole document is only
assembled where every service meets; `migration` is mounted on the entry
point again. The command specs move to the `apps/authup` suite, the emitted
document's shape is pinned in `packages/server-config`, and three specs
still importing the deleted config read modules now read them from
`@authup/server-config`.

BREAKING CHANGE: `defineCLIConfigCommand` is no longer exported by
`@authup/server-core`, and the package no longer ships
`dist/config-schema.json`.
