+++ .agents/structure.md
@@ -7,7 +7,7 @@ It follows hexagonal architecture principles, separating core business logic, ad
 
 | Name                                      | Type        | Description                                                                                           |
 |-------------------------------------------|-------------|-------------------------------------------------------------------------------------------------------|
-| [authup](../apps/authup)                  | CLI         | The operator CLI and the one binary an ordinary deployment runs. It imports the service packages and runs them **in process**: `start`, `core`, `worker`, `migration`, `healthcheck` and `config` (`validate` / `schema`, plan 101 C-2) compose the `defineCLI*Command` helpers server-core exports, so the command bodies stay with the service and a downstream embedder can reuse them; `console [admin\|account\|auth]` is the CLI's own (`src/roles/console.ts`), since a console service is a listener rather than a module graph. No spawn, no supervisor, no config forcing (plan 101 D1). The process the operator starts IS the server, so signals, the exit code and the environment reach it directly, and `PORT`/`HOST` follow the ordinary precedence (environment beats the configuration file, one `authup.yml`). **It is also the only place that knows about all four config registries**: `src/roles/config.ts` holds server-core's plus the three console services', hands each service its own resolved section (`readConsoleConfigs`, resolving path keys against server-core's `rootPath` so one document means the same thing to every service it configures), and passes the console registries into `config validate` / `config schema` so a console's key is neither reported as unread nor missing from the emitted JSON Schema. Root args `--configDirectory` / `--configFile` reach every subcommand, nested ones included (citty runs the root `setup` before it recurses); a stray positional on `core` / `start` / `worker` (the retired `authup start server.core` selector shape) is refused. |
+| [authup](../apps/authup)                  | CLI         | The operator CLI and the one binary an ordinary deployment runs. It imports the service packages and runs them **in process**: `start`, `core`, `worker`, `migration` and `healthcheck` compose the `defineCLI*Command` helpers server-core exports, so those command bodies stay with the service and a downstream embedder can reuse them; `console [admin\|account\|auth]` (`src/commands/console.ts`) and `config` (`validate` / `schema`, `src/commands/config.ts`) are the CLI's own, the first because a console service is a listener rather than a module graph, the second because both subcommands are about the WHOLE `authup.yml` and no single service holds it. No spawn, no supervisor, no config forcing (plan 101 D1). The process the operator starts IS the server, so signals, the exit code and the environment reach it directly, and `PORT`/`HOST` follow the ordinary precedence (environment beats the configuration file, one `authup.yml`). **It is also the only place that knows about all four config registries**: `src/roles/config.ts` holds server-core's plus the three console services', hands each service its own resolved section (`readConsoleConfigs`, resolving path keys against server-core's `rootPath` so one document means the same thing to every service it configures), and `config validate` / `config schema` run against `@authup/server-config`'s whole-document registry, so a console's key is neither reported as unread nor missing from the emitted JSON Schema (which the docs workflow publishes as `docs/src/public/schema/config.json`, generated per deploy rather than committed). Root args `--configDirectory` / `--configFile` reach every subcommand, nested ones included (citty runs the root `setup` before it recurses); a stray positional on `core` / `start` / `worker` (the retired `authup start server.core` selector shape) is refused. |
 | [client-account-console](../apps/client-account-console) | Application | The account console BUNDLE: a client-only Vite/Vue SPA (no SSR, since auth-gated content cannot server-render) for end-user self-service (profile, password, authenticators, connected accounts, sessions, applications). `@authup/server-account-console` depends on the package and serves its built `dist/` with per-request runtime-config injection; the same dist is hostable standalone on any static host (see architecture.md → *Account Console*). No binary, no process, a static bundle. Accepts an optional `ref` query parameter naming the application the visitor came from (validated by the serving service against the trusted app origins), rendered as a back link and carried across every route and the login round-trip. |
 | [client-admin-console](../apps/client-admin-console)          | Application | The admin console BUNDLE: a client-only Vite/Vue SPA (plan 081; no Nuxt, no server process, no binary) in the account-console shape. `@authup/server-admin-console` depends on the package and serves its built `dist/` with per-request runtime-config injection (`<!--admin-config-->` marker -> `window.__AUTHUP__`), authenticating it with the opaque `HttpOnly` session credential (plan 088 Stage 2, whose two server-side routes stay on the API: `GET /console/admin/login/start` and `GET /console/admin/callback`); the same dist is hostable standalone on any static host, where it falls back to the browser-side auth-code + PKCE flow. The former Nuxt `pages/` tree lives on under `src/pages/` unchanged in layout: `src/router.ts` is the explicit route table carrying every former `definePageMeta` as route `meta` (`requireLoggedIn` / `requireLoggedOut` / `requirePermissions` / `layout`), `src/guard.ts` the `router.beforeEach` port of `client-web-nuxt`'s `RoutingInterceptor` (plus the cookie-mode rules), `src/App.vue` a `<Suspense>` around the layout switch so the 12 `async setup()` detail pages stay as they were. Auth entry pages (`/login`, `/login/callback`, `/logout`) carry `meta.layout: 'auth'` and render in the chrome-less `src/layouts/auth.vue`. Self-service lives in the account console: `/settings/*` is a redirect stub mapping each retired path onto its account console route and attaching this console's location as `ref`. |
 | [client-auth-console](../apps/client-auth-console) | Application | The auth console BUNDLE: the Vite-built Vue SSR app behind the hosted auth workflow pages (`/authorize` login+consent, `/register`, `/activate`, `/password-forgot`, `/password-reset`, `/logout`, plus the federated callback's custom-scheme interstitial at `/identity-providers/:id/authorize-in`, which nothing renders since plan 094). `@authup/server-auth-console` depends on the package, renders each request through its built `dist/server/server.js` (per-request hydration payload; client assets under `dist/client/`, mounted by that service at its own `/assets`), and resolves it through the same node_modules ancestor walk the static consoles use (plan 083). The render contract (`src/contract.ts`: `render(RenderContext) => RenderResult`) is the supported customization seam: substitute the package to reskin the hosted auth UI instead of forking anything. It reads `config.basePath` (where the SERVICE is published) separately from `config.baseURL` (the API), which were one value while server-core rendered these pages itself. Architecturally inseparable from the IdP origin (plan 078), so never hosted standalone; unlike its console siblings it has NO per-realm OAuth2 client row, because it IS the IdP surface (see conventions.md → *Workspace Naming*). No binary, no process. |
