+++ .gitignore
@@ -35,3 +35,6 @@ authup.*.conf
 coverage
 .agents/plans
 ds-bundle/
+
+# generated by the docs workflow (`authup config schema`)
+docs/src/public/schema/config.json
