+++ apps/authup/src/commands/config.ts
@@ -5,21 +5,97 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
+import type { AuthupConfig, ConfigReadFsOptions } from '@authup/server-config';
 import {
     CONFIG_SCHEMA,
+    inspectConfigFile,
+    readConfigFileTree,
 } from '@authup/server-config';
 import {
     buildSchemaJSONSchema,
+    mountSchema,
+    readSchemaFromEnv,
+    readSchemaFromFileTree,
 } from '@authup/server-config-kit';
+import { describeConfigError, readConfig } from '@authup/server-core';
 import { defineCommand } from 'citty';
+import process from 'node:process';
+import { Container } from 'validup';
 
-export function defineCLIConfigCommand() {
+/**
+ * Check every key of the document against the zod type its one declaration
+ * carries, not just the ones a single service reads: an operator writes one
+ * file for the whole deployment, so a value a console service will reject has
+ * to be reported here.
+ */
+async function validateDocument(input: Record<string, unknown>) : Promise<void> {
+    const container = new Container<Record<string, unknown>>();
+    mountSchema(container, CONFIG_SCHEMA);
+
+    await container.run(input);
+}
+
+export function defineCLIConfigCommand(
+    configFs: ConfigReadFsOptions<AuthupConfig> = {},
+) {
     return defineCommand({
         meta: {
             name: 'config',
-            description: 'Inspect the configuration the service would read.',
+            description: 'Inspect the configuration the deployment would read.',
         },
         subCommands: {
+            validate: defineCommand({
+                meta: {
+                    name: 'validate',
+                    description: 'Read the configuration file and the environment, and report what does not hold.',
+                },
+                async run() {
+                    try {
+                        // The file itself first: what it holds that nothing
+                        // reads, and whether it was found at all. The read
+                        // below reports neither, because it skips what it does
+                        // not claim and an absent file is a valid deployment.
+                        const { files, unknown } = await inspectConfigFile({ ...configFs });
+
+                        if (files.length === 0 && (configFs.cwd || configFs.file)) {
+                            // Named a place and nothing was there: reporting
+                            // the environment as valid would answer a question
+                            // the operator did not ask.
+                            // eslint-disable-next-line no-console
+                            console.error(`No configuration file was found in ${configFs.cwd || process.cwd()}.`);
+                            process.exit(1);
+                        }
+
+                        if (unknown.length > 0) {
+                            const paths = unknown.map((entry) => `  ${entry}`).join('\n');
+
+                            // eslint-disable-next-line no-console
+                            console.error(`The configuration file holds options that are not read.\n${paths}`);
+                            process.exit(1);
+                        }
+
+                        // server-core's own read next, because normalizing it
+                        // is where the cross-section invariants live (a
+                        // console url on another origin, a throttle without an
+                        // event log). The document pass below carries none of
+                        // them: a zod type only ever sees its own key.
+                        await readConfig({ env: true, fs: { ...configFs } });
+
+                        // and then every key of the document, including the
+                        // ones only a console service reads. The read above
+                        // covers server-core's selection alone.
+                        const { tree } = await readConfigFileTree({ ...configFs });
+                        await validateDocument({
+                            ...readSchemaFromFileTree(tree, CONFIG_SCHEMA),
+                            ...readSchemaFromEnv(CONFIG_SCHEMA),
+                        });
+                    } catch (e) {
+                        // eslint-disable-next-line no-console
+                        console.error(describeConfigError(e));
+                        process.exit(1);
+                    }
+                },
+            }),
             schema: defineCommand({
                 meta: {
                     name: 'schema',
@@ -31,6 +107,9 @@ export function defineCLIConfigCommand() {
                         { title: 'Authup configuration' },
                     );
 
+                    // The docs workflow redirects this stream into
+                    // docs/src/public/schema/config.json, so the document has
+                    // to be the only thing on stdout.
                     // eslint-disable-next-line no-console
                     console.log(JSON.stringify(schema, null, 4));
                 },
