+++ apps/server-core/package.json
@@ -17,9 +17,8 @@
         "cli-dev-dist": "nodemon --watch dist --exec \"node dist/cli/index.mjs\"",
         "build:server:types": "tsc --emitDeclarationOnly -p tsconfig.build.json",
         "build:server:js": "tsdown",
-        "build:server": "npm run build:server:js && npm run build:server:types && npm run build:swagger && npm run build:config-schema",
+        "build:server": "npm run build:server:js && npm run build:server:types && npm run build:swagger",
         "build:swagger": "trapi generate",
-        "build:config-schema": "node scripts/emit-config-schema.mjs",
         "build": "rimraf ./dist && npm run build:server",
         "test": "cross-env NODE_ENV=test vitest run --config test/vitest.config.ts",
         "test:mysql": "cross-env DB_TYPE=mysql DB_HOST=127.0.0.1 DB_USERNAME=root DB_PASSWORD=start123 DB_DATABASE=app vitest run --config test/vitest.config.ts",
